# apps/backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Cidadon API
  version: 0.2.0
  description: API contract for the Cidadon civic demand platform.
servers:
  - url: http://localhost:8080
...
paths:
...
  /auth/register/citizen:
    post:
      summary: Register a citizen
      operationId: registerCitizen
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterCitizenRequest"
      responses:
        "201":
          description: Citizen registered successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterCitizenResponse"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /auth/register/councillor:
    post:
      summary: Register a councillor
      operationId: registerCouncillor
      security: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/RegisterCouncillorMultipartRequest"
      responses:
        "201":
          description: Councillor registered successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterCouncillorResponse"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /auth/register/office-member:
    post:
      summary: Register an office member from an invitation token
      operationId: registerOfficeMember
      security: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/RegisterOfficeMemberMultipartRequest"
      responses:
        "201":
          description: Office member registered successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterOfficeMemberResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /auth/register/office-member/invitation:
    get:
      summary: Preview the valid office invitation associated with a token
      operationId: previewOfficeMemberInvitation
      security: []
      parameters:
        - name: token
          in: query
          required: true
          schema:
            type: string
            minLength: 32
            maxLength: 256
      responses:
        "200":
          description: Public office information for a valid invitation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfficeMemberInvitation"
        "404":
          $ref: "#/components/responses/Error"
...
  /notifications:
    get:
      summary: List persistent notifications for the current user
      operationId: listNotifications
      security: [{ cookieAuth: [] }]
      responses:
        "200": { description: Notifications }
        "401": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
...
  responses:
    Error:
      description: Request failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    Email:
      type: string
      format: email
    Password:
      type: string
      minLength: 6
      maxLength: 72
      writeOnly: true
...
    RegisterBaseRequest:
      type: object
      required:
        - name
        - email
        - password
      properties:
        name:
          type: string
          minLength: 1
        email:
          $ref: "#/components/schemas/Email"
        password:
          $ref: "#/components/schemas/Password"
    RegisterBaseResponse:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
        email:
          $ref: "#/components/schemas/Email"
    RegisterCitizenRequest:
      allOf:
        - $ref: "#/components/schemas/RegisterBaseRequest"
        - type: object
          required:
            - city
            - state
          properties:
            city:
              type: string
              minLength: 1
            state:
              type: string
              minLength: 2
              maxLength: 2
    RegisterCitizenResponse:
      allOf:
        - $ref: "#/components/schemas/RegisterBaseResponse"
        - type: object
          required:
            - city
            - state
          properties:
            city:
              type: string
            state:
              type: string
...
    RegisterCouncillorResponse:
      allOf:
        - $ref: "#/components/schemas/RegisterBaseResponse"
        - type: object
          required:
            - image_url
            - party
            - city
            - state
          properties:
            image_url:
              type: string
            party:
              type: string
            city:
              type: string
            state:
              type: string
    RegisterCouncillorMultipartRequest:
      type: object
      required: [name, email, password, party, city, state, photo]
      properties:
        name: { type: string }
        email: { $ref: "#/components/schemas/Email" }
        password: { $ref: "#/components/schemas/Password" }
        party: { type: string, description: Official party acronym from GET /parties. }
        city: { type: string }
        state: { type: string, minLength: 2, maxLength: 2 }
        photo: { type: string, format: binary, description: Cropped profile photo. }
    RegisterOfficeMemberMultipartRequest:
      type: object
      required: [name, password, token, photo]
      properties:
        name: { type: string }
        password: { $ref: "#/components/schemas/Password" }
        token: { type: string }
        photo: { type: string, format: binary, description: Cropped profile photo. }
    RegisterOfficeMemberResponse:
      allOf:
        - $ref: "#/components/schemas/RegisterBaseResponse"
        - type: object
          required:
            - office_id
            - image_url
          properties:
            office_id:
              type: integer
              minimum: 1
            image_url:
              type: string
    OfficeMemberInvitation:
      type: object
      required:
        - office_id
        - councillor_name
        - party
        - image_url
        - city
        - state
        - expires_at
      properties:
        office_id:
          type: integer
          minimum: 1
        councillor_name:
          type: string
        party:
          type: string
        image_url:
          type: string
          format: uri
        city:
          type: string
        state:
          type: string
          minLength: 2
          maxLength: 2
        expires_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Stable application error code. Frontends should map this code to their own user-facing message.
        message:
          type: string
        details:
          description: Optional structured metadata intended for programmatic clients.
...